Refactor performance counter output formatting. Changed from using PRIu64 to %llu for cycle counter display to ensure proper formatting of large integers.

+++ src/performance.c
@@ -1,7 +1,7 @@
 // Part of PiTubeDirect
 // https://github.com/hoglet67/PiTubeDirect
 #include <stdio.h>
-#include <inttypes.h>
+#include <stdint.h>
 #include <string.h>
 #include "startup.h"
 #include "performance.h"
@@ -223,7 +223,7 @@ void print_performance_counters(perf_counters_t *pct) {
 	int i;
 	uint64_t cycle_counter = pct->cycle_counter;
 	cycle_counter *= 64;
-	printf("%26s = %"PRIu64"\r\n", "cycle counter", cycle_counter);
+	printf("%26s = %llu\r\n", "cycle counter", (unsigned long long)cycle_counter);
 	for (i = 0; i < pct->num_counters; i++) {
 		printf("%26s = %u\r\n", type_lookup(pct->type[i]), pct->counter[i]);
 	}
